+++ scripts/predict/predict_utils.py
@@ -6,17 +6,18 @@ import os
 def split_file(src, cutoff):
     test_fn = os.path.join('exp_files', os.path.basename(src).split(".")[0] + "_test_%d.txt" % (cutoff*100))
     train_fn = os.path.join('exp_files',
-        os.path.basename(src).split(".")[0] + "_train_%d.txt" % ((1-cutoff)*100))
+        os.path.basename(src).split(".")[0] + "_train_%d.txt" % (int((1-cutoff)*100)))
 
-    with open(src, 'rb') as src_file, open(test_fn, 'w') as test_file, \
-            open(train_fn, 'w') as train_file:
+    if not os.path.isfile(test_fn) or not os.path.isfile(train_fn):
+        with open(src, 'rb') as src_file, open(test_fn, 'w') as test_file, \
+                open(train_fn, 'w') as train_file:
 
-        for line in src_file:
-            r = np.random.uniform()
-            if r >= cutoff:
-                train_file.writelines(line)
-            else:
-                test_file.writelines(line)
+            for line in src_file:
+                r = np.random.uniform()
+                if r >= cutoff:
+                    train_file.writelines(line)
+                else:
+                    test_file.writelines(line)
 
     return test_fn, train_fn
